pyproject.toml
[tool.poetry]
name = "decat"
version = "0.1.0"
description = "Declarative linux package manager wrapper"
authors = ["PolyCat <contact.polycat@proton.me>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
